python/calc/requirements.txt
async-timeout==4.0.2
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
flake8==6.0.0
idna==3.4
mccabe==0.7.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyYAML==6.0
redis==4.4.0
requests==2.28.1
tomli==2.0.1
urllib3==1.26.13
